# docs/swagger.yaml
basePath: /api
definitions:
...
  api.EmptyResponse:
    properties:
      data: {}
      meta: {}
    type: object
  api.ErrorDetail:
    properties:
      message:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.ErrorDetail'
    type: object
...
  api.PaginationMeta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  api.TournamentListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Tournament'
        type: array
      meta:
        $ref: '#/definitions/api.PaginationMeta'
    type: object
  api.TournamentRegistrationListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.TournamentRegistration'
        type: array
      meta:
        $ref: '#/definitions/api.PaginationMeta'
    type: object
  api.TournamentRegistrationResponse:
    properties:
      data:
        $ref: '#/definitions/models.TournamentRegistration'
      meta: {}
    type: object
  api.TournamentResponse:
    properties:
      data:
        $ref: '#/definitions/models.Tournament'
      meta: {}
    type: object
...
  api.tournamentRegistrationRequest:
    properties:
      is_invited:
        type: boolean
      manager_contact:
        type: string
      roster_snapshot:
        type: object
      seed_number:
        type: integer
      status:
        type: string
      team_id:
        type: integer
      tournament_id:
        type: integer
    required:
    - team_id
    - tournament_id
    type: object
  api.tournamentRequest:
    properties:
      bracket_config:
        type: object
      currency:
        type: string
      discipline_id:
        type: integer
      end_date:
        type: string
      is_online:
        type: boolean
      name:
        type: string
      prize_pool:
        type: number
      start_date:
        type: string
      status:
        type: string
    required:
    - discipline_id
    - end_date
    - name
    - start_date
    type: object
...
  models.Tournament:
    properties:
      bracket_config:
        type: object
      currency:
        type: string
      discipline_id:
        type: integer
      end_date:
        type: string
      id:
        type: integer
      is_online:
        type: boolean
      name:
        type: string
      prize_pool:
        type: number
      start_date:
        type: string
      status:
        type: string
    type: object
  models.TournamentRegistration:
    properties:
      id:
        type: integer
      is_invited:
        type: boolean
      manager_contact:
        type: string
      registered_at:
        type: string
      roster_snapshot:
        type: object
      seed_number:
        type: integer
      status:
        type: string
      team_id:
        type: integer
      tournament_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: REST API for disciplines, teams, tournaments, matches, and reports.
  title: DB Course Project API
  version: "1.0"
paths:
...
  /tournament-registrations:
    get:
      parameters:
      - description: Tournament ID
        in: query
        name: tournament_id
        type: integer
      - description: Team ID
        in: query
        name: team_id
        type: integer
      - description: Status
        in: query
        name: status
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TournamentRegistrationListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List tournament registrations
      tags:
      - TournamentRegistrations
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.tournamentRegistrationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.TournamentRegistrationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create tournament registration
      tags:
      - TournamentRegistrations
  /tournament-registrations/{id}:
    delete:
      parameters:
      - description: Registration ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/api.EmptyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Delete tournament registration
      tags:
      - TournamentRegistrations
    get:
      parameters:
      - description: Registration ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TournamentRegistrationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get tournament registration
      tags:
      - TournamentRegistrations
    put:
      consumes:
      - application/json
      parameters:
      - description: Registration ID
        in: path
        name: id
        required: true
        type: integer
      - description: Registration payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.tournamentRegistrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TournamentRegistrationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Update tournament registration
      tags:
      - TournamentRegistrations
  /tournaments:
    get:
      parameters:
      - description: Search by name
        in: query
        name: search
        type: string
      - description: Discipline ID
        in: query
        name: discipline_id
        type: integer
      - description: Tournament status
        in: query
        name: status
        type: string
      - description: Start date from (YYYY-MM-DD)
        in: query
        name: start_from
        type: string
      - description: Start date to (YYYY-MM-DD)
        in: query
        name: start_to
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TournamentListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List tournaments
      tags:
      - Tournaments
    post:
      consumes:
      - application/json
      parameters:
      - description: Tournament payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.tournamentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.TournamentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create tournament
      tags:
      - Tournaments
  /tournaments/{id}:
    delete:
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/api.EmptyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Delete tournament
      tags:
      - Tournaments
    get:
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TournamentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get tournament
      tags:
      - Tournaments
    put:
      consumes:
      - application/json
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: integer
      - description: Tournament payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.tournamentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TournamentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Update tournament
      tags:
      - Tournaments
swagger: "2.0"